Move `fit_variogram` from `variogram.py` to `variogram_models.py`
