+++ echopop/nwfsc_feat/variogram.py
@@ -435,133 +435,3 @@ def empirical_variogram(
         )
     else:
         return lags, gamma_h, lag_counts, lag_covariance
-
-
-def fit_variogram(
-    lags: np.ndarray[float],
-    lag_counts: np.ndarray[int],
-    gamma: np.ndarray[float],
-    variogram_parameters: Parameters,
-    model: Union[str, List[str]] = ["bessel", "exponential"],
-    optimizer_kwargs: Dict[str, Any] = {},
-) -> Tuple[Dict[str, Any], float, float]:
-    """
-    Compute the best-fit variogram parameters for input data
-
-    Parameters
-    ----------
-    lags : np.ndarray[float]
-        A 1D array of the lag distances.
-    lag_counts : np.ndarray[int]
-        A 1D array of the lag counts.
-    gamma : np.ndarray[float]
-        A 1D array comprising the semivariance estimates for each lag.
-    variogram_parameters : Parameters
-        A :class:`lmfit.parameter.Parameters` object containing the parameters required for the
-        defined variogram model. See :func:`echopop.spatial.variogram.variogram` for more details.
-        Valid parameters include:
-
-            - **sill : float**
-            The asymptotic value as lags approach infinity
-
-            - **nugget : float**
-            The semivariogram *y*-intercept that corresponds to variability at lag distances
-            shorter than the lag resolution
-
-            - **correlation_range : float**
-            The relative length scale, or range, at which the autocorrelation between lag distances
-            no longer increases and becomes asymptotic
-
-            - **hole_effect_range : float**
-            The (normalized) length scale/range that holes' are observed, which represent 'null'
-            (or very small) points compared to their neighboring lags
-
-            - **decay_power : float**
-            An exponential term that is used in certain generalized exponential (or related)
-            semivariogram models that modulates the ascending rate for a semivariogram
-
-            - **enhance_semivariance : bool**
-            A boolean term that determines whether the correlation decay in certain  cosine-related
-            variogram models are enhanced (or not) with increasing lag distances
-
-    model : Union[str, List[str]], default=['bessel', 'exponential']
-        A string or list of model names. A single name represents a single family model. Two inputs
-        represent the desired composite model (e.g. the composite J-Bessel and exponential model).
-        Defaults to: ``model=["bessel", "exponential"]``. Available models and their required
-        arguments can be reviewed in the :func:`echopop.spatial.variogram.variogram` function.
-    optimizer_kwargs : Dict[str, Any], default={}
-        A dictionary comprising the various function arguments used by
-        :class:`lmfit.minimizer.Minimizer` for Least-Squares minimization that incorporates the
-        Trust Reflective method. See :class:`lmfit.minimizer.Minimizer` for more details.
-
-    Returns
-    -------
-    Tuple[Dict[str, Any], float, float]
-        A tuple containing a dictionary with the best-fit keyword variogram parameter values,
-        the mean absolute deviation (MAD) of the initial parameter values, and the MAD of the
-        best-fit parameter values.
-
-    See Also
-    --------
-    :func:`echopop.spatial.variogram.variogram` :
-        Variogram model parameters.
-    :class:`lmfit.parameter.Parameters` :
-        Variogram parameter optimization leverages the ``Parameters`` class from ``lmfit`` for
-        model optimization.
-    :class:`lmfit.minimizer.Minimizer` :
-        Optimization keyword arguments used for the Least-Squares fitting.
-    """
-    # Normalize the lag counts to get the lag weights
-    lag_weights = lag_counts / lag_counts.sum()
-
-    # Vertically stack the lags, semivariance, and weights
-    data_stack = np.vstack((lags, gamma, lag_weights))
-
-    # Recover the lag resolution
-    delta_lag = np.diff(lags).mean()
-
-    # Compute the range
-    range = lags.max() + delta_lag
-
-    # Index lag distances that are within the parameterized range
-    within_range = np.where(lags <= range)[0]
-
-    # Truncate the data stack
-    truncated_stack = data_stack[:, within_range]
-
-    # Create helper cost-function that is weighted using the kriging weights (`w`), lag
-    # distances (`x`), and empirical semivariance (`y`)
-    def cost_function(parameters, x, y, w, model):
-        yr = variogram(x, {**parameters, **{"model": model}})
-        return (yr - y) * w
-
-    # Compute the initial fit based on the pre-optimized parameter values
-    initial_fit = cost_function(
-        variogram_parameters,
-        x=truncated_stack[0],
-        y=truncated_stack[1],
-        w=truncated_stack[2],
-        model=model,
-    )
-
-    # Compute the initial mean absolute deviation (MAD)
-    mad_initial = np.mean(np.abs(initial_fit))
-
-    # Generate `Minimizer` function class required for bounded optimization
-    minimizer = Minimizer(
-        cost_function,
-        variogram_parameters,
-        fcn_args=(truncated_stack[0], truncated_stack[1], truncated_stack[2], model),
-    )
-
-    # Minimize the cost-function to compute the best-fit/optimized variogram parameters
-    parameters_optimized = minimizer.minimize(method="least_squares", **optimizer_kwargs)
-
-    # Calculate the optimized MAD
-    mad_optimized = np.mean(np.abs(parameters_optimized.residual))
-
-    # Extract the best-fit parameter values
-    best_fit_params = parameters_optimized.params.valuesdict()
-
-    # Return the final tuple
-    return best_fit_params, mad_initial, mad_optimized
